MEMORY_AGENT_SYSTEM/requirements_rag.txt
streamlit>=1.28.0
sqlite-vec>=0.1.6
openai>=1.3.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
streamlit-feedback>=0.1.0
pandas>=2.0.0
numpy>=1.24.0
